Switch OC.CZ connector to offline mode
