+++ main/src/main/java/cgeo/geocaching/connector/oc/OCConnector.java
@@ -1,7 +1,6 @@
 package cgeo.geocaching.connector.oc;
 
 import cgeo.geocaching.R;
-import cgeo.geocaching.connector.AbstractConnector;
 import cgeo.geocaching.connector.UserAction;
 import cgeo.geocaching.connector.capability.Smiley;
 import cgeo.geocaching.connector.capability.SmileyCapability;
@@ -12,8 +11,6 @@ import cgeo.geocaching.network.Network;
 import cgeo.geocaching.utils.ShareUtils;
 import cgeo.geocaching.utils.TextUtils;
 
-import android.net.Uri;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
@@ -23,59 +20,16 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.math.NumberUtils;
-import org.jetbrains.annotations.NotNull;
 
-public class OCConnector extends AbstractConnector implements SmileyCapability {
+public class OCConnector extends OCBaseConnector implements SmileyCapability {
 
-    @NonNull
-    private final String host;
-    private final boolean https;
-    @NonNull
-    private final String name;
-    private final Pattern codePattern;
-    private final String[] sqlLikeExpressions;
     private static final Pattern GPX_ZIP_FILE_PATTERN = Pattern.compile("oc[a-z]{2,3}\\d{5,}\\.zip", Pattern.CASE_INSENSITIVE);
 
     private static final List<LogType> STANDARD_LOG_TYPES = Arrays.asList(LogType.FOUND_IT, LogType.DIDNT_FIND_IT, LogType.NOTE);
     private static final List<LogType> EVENT_LOG_TYPES = Arrays.asList(LogType.WILL_ATTEND, LogType.ATTENDED, LogType.NOTE);
-    @NonNull private final String abbreviation;
 
     public OCConnector(@NonNull final String name, @NonNull final String host, final boolean https, final String prefix, @NonNull final String abbreviation) {
-        this.name = name;
-        this.host = host;
-        this.https = https;
-        this.abbreviation = abbreviation;
-        codePattern = Pattern.compile(prefix + "[A-Z0-9]+", Pattern.CASE_INSENSITIVE);
-        sqlLikeExpressions = new String[]{prefix + "%"};
-    }
-
-    @Override
-    public boolean canHandle(@NonNull final String geocode) {
-        return codePattern.matcher(geocode).matches();
-    }
-
-    @NotNull
-    @Override
-    public String[] getGeocodeSqlLikeExpressions() {
-        return sqlLikeExpressions;
-    }
-
-    @Override
-    @NonNull
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    @NonNull
-    public String getNameAbbreviated() {
-        return abbreviation;
-    }
-
-    @Override
-    @NonNull
-    public String getCacheUrl(@NonNull final Geocache cache) {
-        return getCacheUrlPrefix() + cache.getGeocode();
+        super(name, host, https, prefix, abbreviation);
     }
 
     @Override
@@ -103,45 +57,8 @@ public class OCConnector extends AbstractConnector implements SmileyCapability {
     }
 
     @Override
-    @NonNull
-    public String getHost() {
-        return host;
-    }
-
-    @Override
     public boolean isZippedGPXFile(@NonNull final String fileName) {
         return GPX_ZIP_FILE_PATTERN.matcher(fileName).matches();
-    }
-
-    @Override
-    public boolean isOwner(@NonNull final Geocache cache) {
-        return false;
-    }
-
-    @Override
-    @NonNull
-    protected String getCacheUrlPrefix() {
-        return getSchemeAndHost() + "/viewcache.php?wp=";
-    }
-
-    @Override
-    public int getCacheMapMarkerId() {
-        return R.drawable.marker_oc;
-    }
-
-    @Override
-    public int getCacheMapMarkerBackgroundId() {
-        return R.drawable.background_oc;
-    }
-
-    @Override
-    public int getCacheMapDotMarkerId() {
-        return R.drawable.dot_marker_oc;
-    }
-
-    @Override
-    public int getCacheMapDotMarkerBackgroundId() {
-        return R.drawable.dot_background_oc;
     }
 
     @Override
@@ -152,31 +69,6 @@ public class OCConnector extends AbstractConnector implements SmileyCapability {
         }
 
         return STANDARD_LOG_TYPES;
-    }
-
-    @Override
-    @Nullable
-    public String getGeocodeFromUrl(@NonNull final String url) {
-        // different opencaching installations have different supported URLs
-
-        // host.tld/geocode
-        final String shortHost = getShortHost();
-        final Uri uri = Uri.parse(url);
-        if (!StringUtils.containsIgnoreCase(uri.getHost(), shortHost)) {
-            return null;
-        }
-        final String path = uri.getPath();
-        if (StringUtils.isBlank(path)) {
-            return null;
-        }
-        final String firstLevel = path.substring(1);
-        if (canHandle(firstLevel)) {
-            return firstLevel;
-        }
-
-        // host.tld/viewcache.php?wp=geocode
-        final String secondLevel = path.startsWith("/viewcache.php") ? uri.getQueryParameter("wp") : "";
-        return (secondLevel != null && canHandle(secondLevel)) ? secondLevel : super.getGeocodeFromUrl(url);
     }
 
     @Override
@@ -201,27 +93,6 @@ public class OCConnector extends AbstractConnector implements SmileyCapability {
     @Nullable
     public Smiley getSmiley(final int id) {
         return OCSmileysProvider.getSmiley(id);
-    }
-
-    @Override
-    public boolean isHttps() {
-        return https;
-    }
-
-    /**
-     * Return the scheme part including the colon and the slashes.
-     *
-     * @return either "https://" or "http://"
-     */
-    protected String getSchemePart() {
-        return https ? "https://" : "http://";
-    }
-
-    /**
-     * Return the scheme part and the host (e.g., "https://opencache.uk").
-     */
-    protected String getSchemeAndHost() {
-        return getSchemePart() + host;
     }
 
     @Override
